@@ -6,10 +6,16 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/node": "^20.10.3",
+    "@typescript-eslint/eslint-plugin": "^6.13.2",
+    "@typescript-eslint/parser": "^6.13.2",
+    "eslint-plugin-node": "^11.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "styled-components": "^6.1.1",
+    "typescript": "^5.3.2",
+    "typescript-eslint": "^0.0.1-alpha.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -37,9 +43,11 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^8.0.1",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "eslint": "^8.55.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-eslint-plugin": "^5.1.1",
     "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-import-helpers": "^1.3.1",
     "eslint-plugin-n": "^15.0.0 || ^16.0.0 ",
