@@ -1,13 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
+
 import Login from './containers/Login';
 import GlobalStyles from './styles/globalStyles';
 
-ReactDOM.render(
+const container = document.getElementById('root')
+const root = createRoot(container)
+
+root.render(
   <>
     <Login />
     <GlobalStyles />  
   </>,
 
-  document.getElementById('root')
-);
+)
